BUG/MEDIUM: sample: Fix adjusting size in word converter
Adjust the size of the sample buffer before we change the "area"
pointer. Otherwise, we end up not changing the size, because the area
pointer is already the same as "start" before we compute the difference
between the two.

This is similar to the change in b28430591d18f7fda5bac2e0ea590b3a34f04601
but for the word converter instead of field.

(cherry picked from commit 6a0d2176281bc59e8d81442a056495eb11c5a051)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>