BUG/MEDIUM: ssl_ckch: Don't delete a cert entry if it is being modified
When a certificate entry is being modified, we must take care to no delete
it because the corresponding ongoing transaction still references it. If we
do so, it leads to a null-deref and a crash may be exeperienced if changes
are commited.

This patch must be backported as far as 2.2.

(cherry picked from commit 926fefca8dba3ca7aed3ef7fccf6aa5cffe3cf05)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 6ed5044b510f051f8fa51753728a5c79320a04f8)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>