BUG/MEDIUM: ssl_ckch: Don't delete CA/CRL entry if it is being modified
When a CA or a CRL entry is being modified, we must take care to no delete
it because the corresponding ongoing transaction still references it. If we
do so, it leads to a null-deref and a crash may be exeperienced if changes
are commited.

This patch must be backported as far as 2.5.

(cherry picked from commit 1f08fa46fb5c2ec02f740b716a9674dc72a5cadc)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>