MEDIUM: http-ana: Always report rewrite failures as PRXCOND in logs
Rewrite failures in http rules are reported as proxy errors (PRXCOND) in
logs. However, other rewrite errors are reported as internal errors. For
instance, it happens when we fail to add X-Forwarded-For header. It is not
consistent and it is confusing. So now, all rewite failures are reported as
proxy errors.

This patch may be backported if necessary.

(cherry picked from commit d649b575192d562fc4cbec008c3a2dfca1ee255e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>