REGTESTS: http_request_buffer: Increase client timeout to wait "slow" clients
The default client timeout is too small to be sure to always wait end of
slow clients (the last 2 clients use a delay to send their request). But it
cannot be increased because it will slow down the regtest execution. So a
dedicated frontend with a higher client timeout has been added. This
frontend is used by "slow" clients. The other one is used for normal
requests.

(cherry picked from commit 33a2745c8738f2b163196d88f23188166a579643)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>