BUG/MINOR: ssl_ckch: Use right type for old entry in show_crlfile_ctx
Because of a typo (I guess), an unknown type is used for the old entry in
show_crlfile_ctx structure. Because this field is unused, there is no
compilation error. But it must be a cafile_entry and not a crlfile_entry.

Note this field is not used for now, but it will be used.

This patch must be backported to 2.6.

(cherry picked from commit 51095ee236bbbdfdcb1e71ef102c0912b79b4ad7)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>