BUG/MINOR: ssl_ckch: Dump CRL transaction only once if show command yield
When loaded CRL files are displayed via "show ssl crl-file" command, the
in-progess transaction, if any, is also displayed. However, if the command
yield, the transaction is re-displayed again and again.

To fix the issue, old_crlfile_entry field is used to remember the transaction
was already displayed.

This patch must be backported as far as 2.5.

(cherry picked from commit 9a99e54787634532d18cd5aca3552eeb48a866bf)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>