BUG/MINOR: ssl_ckch: Dump cert transaction only once if show command yield
When loaded SSL certificates are displayed via "show ssl cert" command, the
in-progess transaction, if any, is also displayed. However, if the command
yield, the transaction is re-displayed again and again.

To fix the issue, old_ckchs field is used to remember the transaction was
already displayed.

This patch must be backported as far as 2.2.

(cherry picked from commit 3e94f5d4b6ee7f19ead9925df080e31b60910ef5)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>