BUG/MEDIUM: dns: Keep the right count of active nameservers for a resolver
nb_nameservers counter is incremented when a new connection to a nameserver
is established. However it is never decremented when a connection error it
detected. Thus, this counter may not reflect the number of active
nameservers.

It is problem because this counter is used to know when all queries was sent
on retries for a given resolution. It is also used to know when all
responses was received. So, now, when an connection error is detected,
nb_nameservers counter is decremented. And it is incremented when a new
connection is openned. This way, it always represent the number of active
nameservers for a resolver.

By fixing this counter, another bug was revealed. On send error, the query
is considered as sent (nb_queries is incremented) but we still wait for the
response corresponding to this query. To fix this issue, in this case, the
number of received responses must also be incremented.

However, because the design is totally buggy, this patch remains a partial
fix. There are still several ways to send the same query several times to
the same nameserver. The main reason is that, on retry, we know some queries
must still be sent but we don't know if it was sent or not for a given
nameservers. Thus with several nameservers and several resolutions, we
cannot predict the result on send retries.

This patch should fix (at least partially) the issue #1688. There is no
reason to backport it. And there is no upstream ID because it is specific to
HAProxy 2.2 and below.