BUG/MINOR: qpack: support header litteral name decoding
Complete QPACK decoding with full implementation of litteral field name
with litteral value representation. This change is mandatory to support
decoding of headers name not present in the QPACK static table.
Previously, these headers were silently ignored and not transferred on
the backend request.

QPACK decoding should now be sufficient to deal with all real
situations. Only post-base indices representation are not handled but
this should not cause a problem as they are only used for the dynamic
table whose size is null as announced by the haproxy implementation.

The direct impact of this change is that it should now be possible to
use complex webapp through a QUIC frontend.

This must be backported up to 2.6.

(cherry picked from commit 4bcaf69dca3552579d7ad9393e90d89969f1135c)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>