MINOR: qpack: improve decoding function
Adjust decoding loop by using temporary ist for header name and value.
The header is inserted at the end of an iteration, which guarantee that
we do not insert name only in the list in case of an error on value
decoding. This also helps the function readability by centralizing the
LIST insert operation.

The return value of the decoding function is also changed. Now on
success the number of headers inserted in the input list is returned.
This change as no impact as success value is not used by the caller.
This is mainly done to have a behavior similar to hpack decoding
function.

(cherry picked from commit b666c6b26eae3f17eee058eb6bcc9fe1b1c304d2)
[wt: was partially merged already as part of 644ef94684a]
Signed-off-by: Willy Tarreau <w@1wt.eu>