MINOR: stream: Rely on stconn flags to abort stream destructive upgrade
On destructive connection upgrade, instead of using the new mux name to
abort the old stream, we can relay on the stream connector flags. If it is
detached after the upgrade, it means the stream will not be resused by the
new mux and it must be aborted.

This patch may be backported to 2.6.

(cherry picked from commit 9b8d7a11c0a8513f27780c3be0e43bb2ed5ee040)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>