BUG/MINOR: quic: Wrong reuse of fulfilled dgram RX buffer
After having fulfilled a buffer, then marked it as full, we must
consume the remaining space. But to do that, and not to erase the
already existing data, we must check there is not remaining data in
after the tail of the buffer (between the tail and the head).
This is done adding a condition to test that adding the number of
bytes from the remaining contiguous space to the tail does not
pass the wrapping postion in the buffer.

Must be backported to 2.6.

(cherry picked from commit 0c535683eef05b420d7bf0b25cf38820d02333fe)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>